【phi】Split device_event_def.h as .cc  (#71136)
* split device file

* fix path

* add setup.py

* split device_event_def.h
