+++ paddle/phi/core/platform/CMakeLists.txt
@@ -22,7 +22,7 @@ if(WITH_CUSTOM_DEVICE)
 endif()
 
 # device event library
-list(APPEND DEVICE_SRCS device_event_base.cc)
+list(APPEND DEVICE_SRCS device_event_base.cc device_event_defs.cc)
 if(WITH_GPU OR WITH_ROCM)
   list(APPEND DEVICE_SRCS device_event_gpu.cc)
 endif()
